+++ src/utility/+mag/+internal/stopIfException.m
@@ -1,6 +1,16 @@
 function stopIfException(identifier)
 % STOPIFEXCEPTION Stop if exception is caught.
 
-    dbstop("if", "error", identifier);
-    dbstop("if", "caught error", identifier);
+    arguments
+        identifier string {mustBeScalarOrEmpty}
+    end
+
+    if isempty(identifier) || (strlength(identifier) == 0)
+        identifier = {};
+    else
+        identifier = {identifier};
+    end
+
+    dbstop("if", "error", identifier{:});
+    dbstop("if", "caught error", identifier{:});
 end