+++ resources/release-notes.md
@@ -1,3 +1,9 @@
+## App
+
+- (IMAP) Add support for loading HK defined in XTCE-decoded CSV files
+- (IMAP) Add enumeration for HK type (`mag.meta.HKType`)
+- (All) Fix issue with "Debug" toolstrip button when error identifier is empty
+
 ## Software
 
 - (All) Replace range as integer/double with an enumeration `mag.meta.Range`